Added "util" package; moved appropriate functions into "util" package
